Sign In Activity UI.
This is just the user interface. There is no code backing it yet.
This commit also imports Timber for loging, and does some basic setup of
it.

As this is the first Acitivty added, there's also a load of
look-and-feel stuff generated by Android Studio that's added for the
first time in this commit too.