Introduce a ChooseTeam activity and UserScope.
This introduces a (currently blank) new activity for choosing the team
to be launched once the user is logged in. Accompanying changes are the
necessary model classes, API classes, etc etc to enable this, as well as
a UserScope dagger2 scope for injecting the user and credentials and
whatnot.

Need to actually implement the functionality of the ChangeTeam activity
to make sure this all works as anticipated.