Refactor all session logic into SessionManager
This creates a new class, SessionManager which can be injected in the
App scope which takes care of managing all of the login and team
choosing process, including persistency of tokens and credentials and
all associated network requests. A Rx based event bus is introduced for
transmitting the responses to those network requsts back to the relevant
activities, and as a result a lot of complexity is removed from the
launch and login activities.