Set team on clicking on one & go to main activity.
Still need to figure out how to do saving of the team in the session so
we don't show the team chooser when the user has already selected one
previously. Going to defer this for now as it's a bit of a fiddle.

Also need to have a think about activity finalising (or not).